% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_possible_words.R
\name{find_possible_words}
\alias{find_possible_words}
\title{Find possible words}
\usage{
find_possible_words(mers_tree, known_words, word_lengths, word_letters)
}
\arguments{
\item{mers_tree}{K-mers within letter matrix.}

\item{known_words}{Table of known words.}

\item{word_lengths}{Lengths of words to find.}

\item{word_letters}{First letters of words to find.}
}
\value{
A character vector of known words that are in the letter matrix
}
\description{
Find possible words in letter matrix.
}
