% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mers_tree.R
\name{get_next_chars}
\alias{get_next_chars}
\title{Get next chars}
\usage{
get_next_chars(focus_char, focus_char_pos, prev_chars, letter_matrix)
}
\arguments{
\item{focus_char}{Your name (character string; e.g. "john doe").}

\item{focus_char_pos}{Your name (character string; e.g. "john doe").}

\item{prev_chars}{Your name (character string; e.g. "john doe").}

\item{letter_matrix}{Your name (character string; e.g. "john doe").}
}
\value{
A character string, capitalized to title case.
}
\description{
Greet a person and appropriately capitalize their name.
}
