% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mers_tree.R
\name{get_path}
\alias{get_path}
\title{Get path}
\usage{
get_path(path_char, next_char)
}
\arguments{
\item{path_char}{Your name (character string; e.g. "john doe").}

\item{next_char}{Your name (character string; e.g. "john doe").}
}
\value{
A character string, capitalized to title case.
}
\description{
Greet a person and appropriately capitalize their name.
}
