% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_known_words.R
\name{get_random_word}
\alias{get_random_word}
\title{Get random word}
\usage{
get_random_word(known_words, word_length)
}
\arguments{
\item{known_words}{Table of known words}

\item{word_length}{Number of characters of word to be randomly drawn}
}
\description{
Randomly draw word from known words.
}
