% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_known_words.R
\name{load_known_words}
\alias{load_known_words}
\title{Load known words}
\usage{
load_known_words(known_words_file, n_words)
}
\arguments{
\item{known_words_file}{Known words file path}

\item{n_words}{Number of words to use}
}
\description{
Load known words from file.
}
