% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_letter_matrix.R
\name{make_letter_matrix}
\alias{make_letter_matrix}
\title{Make letter matrix}
\usage{
make_letter_matrix(word_vector, v_size = 3)
}
\arguments{
\item{word_vector}{Input words}

\item{v_size}{Number of columns in output matrix}
}
\value{
Output matrix comprising input words
}
\description{
Make a letter matrix from input words.
}
