% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mers_tree.R
\name{make_mers_tree}
\alias{make_mers_tree}
\title{Make mers tree}
\usage{
make_mers_tree(letter_matrix)
}
\arguments{
\item{letter_matrix}{Your name (character string; e.g. "john doe").}
}
\value{
A character string, capitalized to title case.
}
\description{
Greet a person and appropriately capitalize their name.
}
