% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mers_tree.R
\name{restrict_using_first_2_letters}
\alias{restrict_using_first_2_letters}
\title{Filter words by k-mers}
\usage{
restrict_using_first_2_letters(x, possible_strings, n_letter_match = 5)
}
\arguments{
\item{x}{Input words}

\item{possible_strings}{Character strings (i.e. k-mers)}

\item{n_letter_match}{K-mer string length}
}
\value{
A vector of words
}
\description{
Filter input words to only words that include the possible character strings (i.e. the k-mers)
}
