% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_solve_random_letter_marix.R
\name{run_make_letter_matrix}
\alias{run_make_letter_matrix}
\title{Make letter matrix from known words}
\usage{
run_make_letter_matrix(known_words, word_length)
}
\arguments{
\item{known_words}{Table of known words}

\item{word_length}{Number of characters in words}
}
\value{
Letter matrix
}
\description{
Make letter matrix from known words.
}
