% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mers_tree.R
\name{subset_known_words}
\alias{subset_known_words}
\title{Subset known words}
\usage{
subset_known_words(x, word_length, first_letters = NA)
}
\arguments{
\item{x}{Your name (character string; e.g. "john doe").}

\item{word_length}{Your name (character string; e.g. "john doe").}

\item{first_letters}{Your name (character string; e.g. "john doe").}
}
\value{
A character string, capitalized to title case.
}
\description{
Subset known words table by word length and first letter(s).
}
