% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\docType{data}
\name{word_freq_file}
\alias{word_freq_file}
\title{Word frequency file path}
\format{
An object of class \code{character} of length 1.
}
\usage{
word_freq_file
}
\description{
Location in package for word frequency file
}
\keyword{datasets}
