% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mers_tree.R
\name{make_mers_tree}
\alias{make_mers_tree}
\title{Make mers tree}
\usage{
make_mers_tree(letter_matrix)
}
\arguments{
\item{letter_matrix}{Matrix comprising letters.}
}
\value{
A list with one element comprising all k-mers of equal length. For instance, one element for 1-mers, and
second for 2-mers.
}
\description{
Find k-mers within the letter matrix from k = 1 to 5.
}
