% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mers_tree.R
\name{restrict_using_first_2_letters}
\alias{restrict_using_first_2_letters}
\title{restrict_using_first_2_letters}
\usage{
restrict_using_first_2_letters(x, possible_strings, n_letter_match = 5)
}
\arguments{
\item{x}{Your name (character string; e.g. "john doe").}

\item{possible_strings}{Your name (character string; e.g. "john doe").}

\item{n_letter_match}{Your name (character string; e.g. "john doe").}
}
\value{
A character string, capitalized to title case.
}
\description{
Greet a person and appropriately capitalize their name.
}
